+++ Interfaces/VideoSubFinderWXW/ImageBox.h
@@ -64,7 +64,7 @@ public:
 	void ResizeControls();
 
 	void ViewRGBImage(simple_buffer<int> &Im, int w, int h);
-	void ViewGrayscaleImage(simple_buffer<int> &Im, int w, int h);
+	void ViewGrayscaleImage(simple_buffer<u8> &Im, int w, int h);
 	void ViewImage(simple_buffer<int> &Im, int w, int h);
 	void ViewBGRImage(simple_buffer<u8>& ImBGR, int w, int h);
 	void ClearScreen();
